refactoring: move basic file operations to the appropriate folder
